Crate rgmsh

Unofficial, opinionated Rust bindings for the Gmsh API.

From the Gmsh website:

Gmsh is a free 3D finite element mesh generator with a built-in CAD engine and post-processor.

Gmsh is copyright (C) 1997-2019 by C. Geuzaine and J.-F. Remacle

The full Gmsh reference manual is available at: http://gmsh.info/doc/texinfo/gmsh.html

Examples

See the examples module.

The Gmsh API

Gmsh has four loosely coupled parts:

Geometry

Mesh

Solvers

Post-processing

Rust API design

Gmsh uses tags to keep track of different components. For example, adding a point to a geometry model will return a PointTag.

Tags are used everywhere in Gmsh, not just for geometry. Other examples are:

Mesh elements,

Post-processing data sets,

Mesh refinement fields.

Modules

err
Idiomatic Rust error handling for the Gmsh API.

examples
Basic to advanced examples.

fltk
Manipulation of the FLTK graphical user interface.

interface
Interface to the low-level gmsh_sys crate.

model
Inspection and manipulation of geometry models.

Macros

add_points
Add points to a geometry model inline.

Structs

GeoModel
An instance of the built-in geometry kernel.

Gmsh
Gmsh context object

OccModel
An instance of the OpenCASCADE geometry kernel.

Enums

GmshError
The error type for all Gmsh API functions.

Type Definitions

GmshResult
Type alias for Result using GmshError.