Module rgmsh::interface::geo

The set of built-in kernel functions.

Functions

add_curve_loop⚠
Add a curve loop (a closed wire) formed by the curves curveTags. curveTags should contain (signed) tags of model enties of dimension 1 forming a closed loop: a negative tag signifies that the underlying curve is considered with reversed orientation. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the curve loop.

add_line⚠
Add a straight line segment between the two points with tags startTag and endTag. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the line.

add_plane_surface⚠
Add a plane surface defined by one or more curve loops wireTags. The first curve loop defines the exterior contour; additional curve loop define holes. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the surface.

add_point⚠
Add a geometrical point in the built-in CAD representation, at coordinates (x, y, z). If meshSize is > 0, add a meshing constraint at that point. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the point. (Note that the point will be added in the current model only after synchronize is called. This behavior holds for all the entities added in the geo module.)

remove_point⚠
Remove the entities dimTags. If recursive is true, remove all the entities on their boundaries, down to dimension 0.

synchronize⚠
Synchronize the built-in CAD representation with the current Gmsh model. This can be called at any time, but since it involves a non trivial amount of processing, the number of synchronization points should normally be minimized.